pyproject.toml
[tool.poetry]
name = "ska-mid-dish-b5dc-proxy"
version = "0.0.1"
description = "Tango proxy to the Band 5 down-converter"
authors = ["Team Karoo"]
license= "BSD-3-Clause"
readme = "README.md"

[tool.poetry.scripts]
B5dcProxy = 'ska_mid_dish_b5dc_proxy.b5dc_proxy:main'

[[tool.poetry.source]]
name = 'skao'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'

[tool.poetry.dependencies]
python = "~3.10"
pytango = "9.5.0"
ska-tango-base = "1.0.0"
numpy = "1.26.2"
ska-mid-dish-dcp-lib = "0.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pylint = "^2.17.4"
pylint-junit = "^0.3.2"
pytest = "^8.2.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^4.1.0"
pytest-forked = "^1.4.0"
pytest-json-report = "^1.4.1"
flake8-docstrings = "^1.7.0"
flake8-rst-docstrings = "^0.3.0"
flake8-pyproject = "^1.2.3"
mypy = "^1.3.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^6.2.1"
docutils = "< 0.20"
sphinx-tabs = "*"
typing_extensions = "*"
sphinx-copybutton = "*"
ska-ser-sphinx-theme = "^0.1.2"
sphinx-autodoc-typehints = "^1.23.0"
sphinxcontrib-plantuml = "^0.30"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = "tests"
markers = [
    "unit",
    "acceptance"
]

[tool.flake8]
docstring-style = "sphinx"
rst-directives = "literalinclude"
rst-roles = ["py:attr", "py:class", "py:const", "py:exc", "py:func", "py:meth", "py:mod"]
enable = "DAR104"
max-complexity = 10

# E203 and W503 conflict with black
ignore = ["E203", "W503"]

[tool.mypy]
mypy_path = "src/:tests/"
show_error_codes = true
strict = true

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
[tool.pylint.messages_control]
disable = [
    "duplicate-code",
    "fixme",
]
enable = ["useless-suppression"]

[tool.pylint.similarities]
min-similarity-lines = 7
ignore-imports = "yes"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
